dev: introduce io utils
