+++ src/evox/problems/neuroevolution/reinforcement_learning/env_pool.py
@@ -5,32 +5,10 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 from jax import jit, random, vmap, lax
-from jax.tree_util import tree_map
 from jax.experimental import io_callback
 
 from evox import Problem, State, jit_class
-
-
-def _x32_func_call(func):
-    def inner_func(*args, **kwargs):
-        return _to_x32_if_needed(func(*args, **kwargs))
-
-    return inner_func
-
-
-def _to_x32_if_needed(values):
-    if jax.config.jax_enable_x64:
-        # we have 64-bit enabled, so nothing to do
-        return values
-
-    def to_x32(value):
-        if value.dtype == np.float64:
-            return value.astype(np.float32)
-        elif value.dtype == np.int64:
-            return value.astype(np.int32)
-        else:
-            return value
-    return tree_map(to_x32, values)
+from evox.utils.io import to_x32_if_needed, x32_func_call
 
 
 @jit_class
@@ -60,8 +38,8 @@ class EnvPool(Problem):
         key, subkey = random.split(state.key)
         seed = random.randint(subkey, (1,), 0, jnp.iinfo(jnp.int32).max)
         io_callback(self.env.seed, None, seed)
-        obs, info = _to_x32_if_needed(self.env.reset(None))
-        obs, info = io_callback(_x32_func_call(self.env.reset), (obs, info), None)
+        obs, info = to_x32_if_needed(self.env.reset(None))
+        obs, info = io_callback(x32_func_call(self.env.reset), (obs, info), None)
         total_reward = 0
         i = 0
 
@@ -75,11 +53,11 @@ class EnvPool(Problem):
         def step(loop_state):
             i, done, total_reward, obs = loop_state
             action = self.batch_policy(pop, obs)
-            obs, reward, terminated, truncated, info = _to_x32_if_needed(
+            obs, reward, terminated, truncated, info = to_x32_if_needed(
                 self.env.step(np.zeros(action.shape))
             )
             obs, reward, terminated, truncated, info = io_callback(
-                _x32_func_call(lambda action: self.env.step(np.copy(action))),
+                x32_func_call(lambda action: self.env.step(np.copy(action))),
                 (obs, reward, terminated, truncated, info),
                 action,
             )
