+++ src/evox/utils/__init__.py
@@ -1,1 +1,2 @@
 from .common import *
+import io
